requirements.txt
appnope==0.1.0
asn1crypto==0.22.0
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
biom-format==2.1.6
biopython==1.70
burrito==0.9.1
burrito-fillings==0.1.1
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cogent==1.5.3
conda==4.3.30
cryptography==2.0.3
cutadapt==1.14
cycler==0.10.0
decorator==4.1.2
emperor==0.9.60
enum34==1.1.6
future==0.16.0
gdata==2.0.18
idna==2.6
ipaddress==1.0.18
ipython==5.5.0
ipython-genutils==0.2.0
matplotlib==2.1.0
matplotlib-venn==0.11.5
natsort==3.5.6
numpy==1.13.3
pandas==0.21.0
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pybedtools==0.7.10
pycosat==0.6.2
pycparser==2.18
Pygments==2.2.0
pynast==1.2.2
pyOpenSSL==17.2.0
pyparsing==2.2.0
pyqi==0.3.2
pysam==0.13
PySocks==1.6.7
python-dateutil==2.6.1
pytz==2017.3
qcli==0.1.1
qiime==1.9.1
qiime-default-reference==0.1.3
requests==2.18.4
ruamel-yaml==0.11.14
ruffus==2.6.3
scandir==1.6
scikit-bio==0.2.3
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
subprocess32==3.2.7
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
xopen==0.2.1
